+++ news-service/src/main/java/ru/clevertec/statkevich/newsservice/service/NewsServiceImpl.java
@@ -5,7 +5,11 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import ru.clevertec.statkevich.newsservice.domain.News;
 import ru.clevertec.statkevich.newsservice.dto.news.NewsUpdateDto;
 import ru.clevertec.statkevich.newsservice.filter.Filter;
@@ -14,6 +18,7 @@ import ru.clevertec.statkevich.newsservice.repository.NewsRepository;
 import ru.clevertec.statkevich.newsservice.service.api.NewsService;
 
 @RequiredArgsConstructor
+@Transactional(readOnly = true)
 @Service
 public class NewsServiceImpl implements NewsService {
 
@@ -21,6 +26,7 @@ public class NewsServiceImpl implements NewsService {
 
     private final NewsMapper newsMapper;
 
+    @Transactional
     @Override
     public News create(News news) {
         newsRepository.saveAndFlush(news);
@@ -48,17 +54,33 @@ public class NewsServiceImpl implements NewsService {
         return newsRepository.findAll(filter, pageable);
     }
 
+    @Transactional
     @Override
     public News update(Long id, NewsUpdateDto newsUpdateDto) {
-
         News news = findById(id);
-        newsMapper.map(newsUpdateDto, news);
-        newsRepository.save(news);
-        return news;
+        if (isNewsOwner(news)) {
+            newsMapper.map(newsUpdateDto, news);
+            newsRepository.save(news);
+            return news;
+        }
+        throw new AccessDeniedException("modifying prohibited");
     }
 
+    @Transactional
     @Override
     public void delete(Long id) {
-        newsRepository.deleteById(id);
+        News news = findById(id);
+        if (isNewsOwner(news)) {
+            newsRepository.deleteById(id);
+        }
+        throw new AccessDeniedException("modifying prohibited");
+    }
+
+    private boolean isNewsOwner(News news) {
+        UserDetails principal = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        String usernamePrincipal = principal.getUsername();
+        String username = news.getUsername();
+
+        return username.equals(usernamePrincipal);
     }
 }
